Trying to get firebase realtime database to work
