+++ ListApp/ListApp/app/src/main/java/com/example/listapp/dataClasses/List.kt
@@ -1,3 +1,3 @@
 package com.example.listapp.dataClasses
 
-data class List(var title: String)
+data class List(var title: String? = null)
