+++ ListApp/ListApp/app/src/main/java/com/example/listapp/service/ToDoListService.kt
@@ -1,0 +1,7 @@
+package com.example.listapp.service
+
+// Firebase code here and send/recieve from data managers
+class ToDoListService {
+
+
+}
